enhance get courses and cleaning code
