+++ src/pages/manager/Courses/hooks.ts
@@ -2,23 +2,26 @@ import { useQuery } from '@tanstack/react-query'
 import { useMemo } from 'react'
 import { getCourses } from '../../../api/services/course-service'
 import { camelCaseKeys } from '../../../utils/formatter'
-import type { IGetCoursesResponse } from './types'
+import type { IGetCoursesResponse } from '../../../utils/global-types'
 
 const useCustom = () => {
   const { data, isLoading } = useQuery({
-    queryKey: ['courseList'],
+    queryKey: ['courses'],
     queryFn: () => getCourses(),
   })
 
-  const courseList = useMemo(
-    () => data?.data.map((item: IGetCoursesResponse) => camelCaseKeys(item)),
+  const courses = useMemo(
+    () =>
+      (data?.data || []).map((item: IGetCoursesResponse) =>
+        camelCaseKeys(item)
+      ),
     [data]
   )
 
   return {
     data: {
       loading: isLoading,
-      courseList,
+      courses,
     },
     methods: {},
   }
