+++ src/utils/global-types.ts
@@ -5,3 +5,19 @@ export interface IStorageKey {
   photo: string
   token: string
 }
+
+export interface IGetCoursesResponse {
+  _id: number
+  title: string
+  thumbnail: string
+  total_students: number
+  category: string
+  students: string[]
+}
+
+export interface IGetCategoriesResponse {
+  _id: string
+  name: string
+  courses: string[]
+  updatedAt: string
+}
